requirements.txt
python-telegram-bot==21.6
gspread==6.1.4
google-auth==2.35.0
python-dotenv==1.0.1
